Move cache/fsutil into a separate cache/util subpackage
